20130403 1714

+++ GP_Tree_Swap.f90
@@ -8,28 +8,28 @@ use GA_Variables
 
 implicit none
 
-real (kind=4) :: cff
+real(kind=8) :: cff
 
-integer (kind=4) :: Child_Tree_Swap_Node_Type(n_Nodes,2)
-integer (kind=4) :: Node_Depth(n_nodes,2)
-integer (kind=4) :: i_Parent,i_Child
-integer (kind=4) :: i_Parent_One,i_Parent_Two
-integer (kind=4) :: i_Level,i_Levels
-integer (kind=4) :: i_Parent_Level,i_Child_Level
-integer (kind=4) :: i_Node_at_Level,n_Nodes_at_Level
-integer (kind=4) :: i_Node,j_Node,k_Node
-integer (kind=4) :: i
-integer (kind=4) :: icnt
-integer (kind=4) :: icff
-integer (kind=4) :: icnt_parent_one_nodes,icnt_parent_two_nodes
-integer (kind=4) :: i_parent_one_swap_node,i_parent_two_swap_node
-integer (kind=4) :: i_parent_swap_node,i_child_swap_node
-integer (kind=4) :: parent_max_swap_level,child_max_swap_level
-integer (kind=4) :: i_parent_node_pt,i_child_node_pt
-integer (kind=4) :: i_level_node
-integer (kind=4) :: n_parent_one_swap_levels,n_parent_two_swap_levels
-integer (kind=4) :: parent_one_max_swap_level,parent_two_max_swap_level
-integer (kind=4) :: parent_two_swappable_nodes(n_nodes)
+integer(kind=4) :: Child_Tree_Swap_Node_Type(n_Nodes,2)
+integer(kind=4) :: Node_Depth(n_nodes,2)
+integer(kind=4) :: i_Parent,i_Child
+integer(kind=4) :: i_Parent_One,i_Parent_Two
+integer(kind=4) :: i_Level,i_Levels
+integer(kind=4) :: i_Parent_Level,i_Child_Level
+integer(kind=4) :: i_Node_at_Level,n_Nodes_at_Level
+integer(kind=4) :: i_Node,j_Node,k_Node
+integer(kind=4) :: i
+integer(kind=4) :: icnt
+integer(kind=4) :: icff
+integer(kind=4) :: icnt_parent_one_nodes,icnt_parent_two_nodes
+integer(kind=4) :: i_parent_one_swap_node,i_parent_two_swap_node
+integer(kind=4) :: i_parent_swap_node,i_child_swap_node
+integer(kind=4) :: parent_max_swap_level,child_max_swap_level
+integer(kind=4) :: i_parent_node_pt,i_child_node_pt
+integer(kind=4) :: i_level_node
+integer(kind=4) :: n_parent_one_swap_levels,n_parent_two_swap_levels
+integer(kind=4) :: parent_one_max_swap_level,parent_two_max_swap_level
+integer(kind=4) :: parent_two_swappable_nodes(n_nodes)
 
 logical node_not_found
 
@@ -44,21 +44,21 @@ i_parent_two=2
 
 ! determine from the bottom of the tree what the summed tree levels are for each node for both parents
 do i=1,2
-  if (i .eq. 1) i_parent=i_parent_one
-  if (i .eq. 2) i_parent=i_parent_two
+  if( i .eq. 1) i_parent=i_parent_one
+  if( i .eq. 2) i_parent=i_parent_two
   do i_level=n_levels,1,-1
     n_nodes_at_level=int(2**(i_level-1))
     do i_node_at_level=1,n_nodes_at_level
       i_node=(n_nodes_at_level-1)+i_node_at_level
       node_depth(i_node,i_parent)=0
-      if (parent_Tree_Swap_Node_Type(i_node,i_parent) .ge. -1) then
-        if (i_level .eq. n_levels) then
+      if( parent_Tree_Swap_Node_Type(i_node,i_parent) .ge. -1) then
+        if( i_level .eq. n_levels) then
           node_depth(i_node,i_parent)=1
         else
           j_node=i_node*2
           k_node=1+(i_node*2)
-          if (node_depth(j_node,i_parent) .ge. 1 .and. node_depth(k_node,i_parent) .ge. 1) then
-            if (node_depth(k_node,i_parent) .gt. node_depth(j_node,i_parent)) then
+          if( node_depth(j_node,i_parent) .ge. 1 .and. node_depth(k_node,i_parent) .ge. 1) then
+            if( node_depth(k_node,i_parent) .gt. node_depth(j_node,i_parent)) then
               j_node=k_node
             endif
             node_depth(i_node,i_parent)=node_depth(j_node,i_parent)+1
@@ -74,7 +74,7 @@ enddo
 ! count the number of nodes on parent_one with code on it.
 icnt_parent_one_nodes=0
 do i_node=1,n_nodes
-  if (Parent_Tree_Swap_Node_Type(i_node,i_parent_one) .ge. -1) then
+  if( Parent_Tree_Swap_Node_Type(i_node,i_parent_one) .ge. -1) then
     icnt_parent_one_nodes=icnt_parent_one_nodes+1
   endif
 enddo
@@ -85,10 +85,10 @@ i_parent_one_swap_node=1+int(float(icnt_parent_one_nodes)*cff)
 icnt=0
 node_not_found=.true.
 do i_node=1,n_nodes
-  if (node_not_found) then
-    if (parent_Tree_Swap_Node_Type(i_node,i_parent_one) .ge. -1) then
+  if( node_not_found) then
+    if( parent_Tree_Swap_Node_Type(i_node,i_parent_one) .ge. -1) then
       icnt=icnt+1
-      if (icnt .eq. i_parent_one_swap_node) then
+      if( icnt .eq. i_parent_one_swap_node) then
         i_parent_one_swap_node=i_node
         node_not_found=.false.
       endif
@@ -102,7 +102,7 @@ n_parent_one_swap_levels=node_depth(i_parent_one_swap_node,i_parent_one)
 parent_one_max_swap_level=1
 do i_level=2,n_levels
   icff=(2**(i_level-1))-1
-  if (i_parent_one_swap_node .gt. icff) then
+  if( i_parent_one_swap_node .gt. icff) then
     parent_one_max_swap_level=i_level
   endif
 enddo
@@ -127,7 +127,7 @@ do i_level=1,n_levels-1
 
   do i_level_node=1,n_nodes_at_level
     i_node=i_node+1
-    if (parent_Tree_Swap_Node_Type(i_node,i_parent_two) .ge. -1 .and. &
+    if( parent_Tree_Swap_Node_Type(i_node,i_parent_two) .ge. -1 .and. &
         node_depth(i_node,i_parent_two) .le. parent_one_max_swap_level .and. &
         n_parent_one_swap_levels .le. parent_two_max_swap_level) then
       icnt_parent_two_nodes=icnt_parent_two_nodes+1
@@ -147,7 +147,7 @@ n_parent_two_swap_levels=node_depth(i_parent_two_swap_node,i_parent_two)
 parent_two_max_swap_level=1
 do i_level=2,n_levels
   icff=(2**(i_level-1))-1
-  if (i_parent_two_swap_node .gt. icff) then
+  if( i_parent_two_swap_node .gt. icff) then
     parent_two_max_swap_level=i_level
   endif
 enddo
@@ -159,7 +159,7 @@ parent_two_max_swap_level=n_levels-parent_two_max_swap_level+1
 
 do i_child=1,2
 
-  if (i_child .eq. 1) then
+  if( i_child .eq. 1) then
     i_parent_swap_node=i_parent_two_swap_node
     i_child_swap_node=i_parent_one_swap_node
     parent_max_swap_level=parent_two_max_swap_level
@@ -171,7 +171,7 @@ do i_child=1,2
     child_max_swap_level=parent_two_max_swap_level
   endif
 
-  if (i_child .eq. 1) then
+  if( i_child .eq. 1) then
     i_parent=i_parent_one
   else
     i_parent=i_parent_two
@@ -179,7 +179,7 @@ do i_child=1,2
   child_Tree_Swap_Node_Type(1:n_Nodes,i_child)=parent_Tree_Swap_Node_Type(1:n_Nodes,i_parent)
 
 ! set each of the children to one of the parents initially
-  if (i_child .eq. 1) then
+  if( i_child .eq. 1) then
     i_parent=i_parent_two
   else
     i_parent=i_parent_one
@@ -198,7 +198,7 @@ do i_child=1,2
     enddo
   enddo
 
-  if (i_child .eq. 1) then
+  if( i_child .eq. 1) then
     i_parent_swap_node=i_parent_two_swap_node
     i_child_swap_node=i_parent_one_swap_node
     parent_max_swap_level=parent_two_max_swap_level
@@ -215,7 +215,7 @@ do i_child=1,2
   do i_Child_Level=n_Levels-Child_Max_Swap_Level+2,n_Levels
     i_Levels=i_Levels+1
     i_Parent_Level=n_Levels-Parent_Max_Swap_Level+1+i_Levels
-    if (i_parent_level .le. n_levels) then
+    if( i_parent_level .le. n_levels) then
       i_child_swap_node=i_child_swap_node*2
       i_parent_swap_node=i_parent_swap_node*2
       do i_node=1,2**i_levels
